deposits: add serializer for deposit `access`
* uses a new serializer for search, that does not include
all the profile info in the `access` field
