+++ cap/modules/deposit/serializers/__init__.py
@@ -30,13 +30,14 @@ from invenio_records_rest.serializers.response import (record_responsify,
                                                        search_responsify)
 
 from .json import DepositSerializer
-from .schemas.json import DepositSchema, DepositFormSchema
+from .schemas.json import DepositSchema, DepositFormSchema, DepositSearchSchema
 
 # Serializers
 # ===========
 # CAP JSON serializer version 1.0.0
 deposit_json_v1 = DepositSerializer(DepositSchema)
 deposit_form_json_v1 = DepositSerializer(DepositFormSchema)
+deposit_search_json_v1 = DepositSerializer(DepositSearchSchema)
 
 # Records-REST serializers
 # ========================
@@ -45,8 +46,10 @@ deposit_json_v1_response = record_responsify(deposit_json_v1,
                                              'application/json')
 deposit_form_json_v1_response = record_responsify(deposit_form_json_v1,
                                                   'application/json')
-deposit_json_v1_search = search_responsify(deposit_json_v1, 'application/json')
+deposit_json_v1_search = search_responsify(deposit_search_json_v1,
+                                           'application/json')
 
 # Files-REST serializers
+# ========================
 # JSON Files serializers for deposit files
 files_response = json_file_response
